java-library plugins works with groovy
At the moment the Java Library plugin is only wired to behave correctly
with the java plugin. Other plugins, such as the Groovy plugin, may not
behave correctly. In particular, if the Groovy plugin is used in
addition to the java-library plugin, then consumers may not get the
Groovy classes when they consume the library. To workaround this, you
need to explicitly wire the Groovy compile dependency, like this: